WIP: Stop recommending imagemagick in the admin pages
This tempts admins into installing insecure packages to make the
warning go away, even going so far as to [impurely modify running
containers](https://github.com/nextcloud/docker/issues/1414#issuecomment-1008915705).

Changing the warning to trigger if the php-imagick module is loaded is
more in line with the actual upstream recommendation, and hopefully
helps unsuspecting users who have hacked around the warning in the
past realize this.

Draft because:

- [ ] Untested (will try building this all this evening)
- [ ] Translations are missing, covered are:
  - [x] English
  - [x] German

Band-aid for #13099, while we wait for a proper solution with libvips
or somesuch.